+++ example/flake.nix
@@ -33,17 +33,25 @@
               '';
 
               # Create .sqlite database from chinook database.
-              sqlite-init.command = ''
-                echo "$(date): Importing Chinook database ($DATAFILE) ..."
-                ${lib.getExe pkgs.sqlite} "$DATAFILE" < ${inputs.chinookDb}/ChinookDatabase/DataSources/Chinook_Sqlite.sql
-                echo "$(date): Done."
-              '';
+              sqlite-init = { name, ... }: {
+                command = pkgs.writeShellApplication {
+                  inherit name;
+                  text = ''
+                    echo "$(date): Importing Chinook database ($DATAFILE) ..."
+                    ${lib.getExe pkgs.sqlite} "$DATAFILE" < ${inputs.chinookDb}/ChinookDatabase/DataSources/Chinook_Sqlite.sql
+                    echo "$(date): Done."
+                  '';
+                };
+              };
 
               # Run sqlite-web on the local chinook database.
-              sqlite-web = {
-                command = ''
-                  ${pkgs.sqlite-web}/bin/sqlite_web --port ${builtins.toString port} "$DATAFILE"
-                '';
+              sqlite-web = { name, ... }: {
+                command = pkgs.writeShellApplication {
+                  inherit name;
+                  text = ''
+                    ${pkgs.sqlite-web}/bin/sqlite_web --port ${builtins.toString port} "$DATAFILE"
+                  '';
+                };
                 # The 'depends_on' will have this process wait until the above one is completed.
                 depends_on."sqlite-init".condition = "process_completed_successfully";
                 readiness_probe.http_get = {