+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     - #18: Add `testScript` option for adding flake checks based on nixosTest library.
 - Fixes
     - #19: Reintroduce the `shell` option so process-compose doesn't rely on user's global bash (which doesn't exist nixosTest runners).
+    - #22: `command` option is no longer wrapped in `writeShellApplication`.
     - #20: Fix definiton of `probe.exec`